Upload field: if upload button is not enabled, don't show the delete button

+++ templates/Includes/UploadField_FileButtons.ss
@@ -8,7 +8,7 @@
 <% end_if %>
 <button class="ss-uploadfield-item-remove ss-ui-button ui-corner-all" title="<%t UploadField.REMOVEINFO 'Remove this file from here, but do not delete it from the file store' %>" data-icon="plug-disconnect-prohibition">
 <%t UploadField.REMOVE 'Remove' %></button>
-<% if $canDelete %>
+<% if $UploadField.canUpload && $canDelete %>
 	<button data-href="$UploadFieldDeleteLink" class="ss-uploadfield-item-delete ss-ui-button ui-corner-all" title="<%t UploadField.DELETEINFO 'Permanently delete this file from the file store' %>" data-icon="minus-circle"><%t UploadField.DELETE 'Delete from files' %></button>
 <% end_if %>
 <% if $UploadField.canAttachExisting %>
